requirements-backup.txt
Flask==1.1.1
MarkupSafe==1.1.1
requests==2.22.0
numpy==1.16.4
matplotlib==3.1.1
setuptools==41.0.0
tensorflow==1.13.2
pylint==2.3.1
autopep8==1.4.4
h5py==2.9.0
pyyaml==5.1.2
scipy==1.3.1
sklearn==0.0
wiotp-sdk==0.9.5
watson-machine-learning-client==1.0.371
keras==2.1.6